Revert "Moved App Todo to the Todo component."
This reverts commit 8d17fb22d4dec49089d4a77dc0e05286e2f49e0e.
